BACKEND_LEGACY/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==5.2.3
django-filter==25.1
djangorestframework==3.16.0
dotenv==0.9.9
google-auth==2.40.3
google-genai==1.19.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Markdown==3.8
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pymongo==3.11.4
pypdfium2==4.30.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
websockets==15.0.1
